refacotr: reorganize project structure into src package layout
