+++ .gitignore
@@ -3,5 +3,7 @@ data/Latest_News.csv
 data/Latest_News.json
 src/.ipynb_checkpoints/
 *.pyc
+__pycache__/
+*.egg-info/
 output/executed_notebooks/
 output/runs/
