+++ pyproject.toml
@@ -1,3 +1,7 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
 [project]
 name = "misinformation-llm-simulation"
 version = "0.1.0"
@@ -19,7 +23,13 @@ dependencies = [
 ]
 
 [tool.uv]
-package = false
+package = true
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
 
 [dependency-groups]
 dev = ["pre-commit", "nbstripout", "ruff"]
