--- src/consts/__init__.py
@@ -1,14 +1,0 @@
-from .datasets import (
-    DETAILED_TEXT_COLUMNS,
-    LANGUAGE_CODE_TO_NAME,
-    TEXT_COLUMN_CANDIDATES,
-)
-from .llm_requests import PROMPT_TEMPLATE, REWRITE_SYSTEM_INSTRUCTION
-
-__all__ = [
-    "TEXT_COLUMN_CANDIDATES",
-    "DETAILED_TEXT_COLUMNS",
-    "LANGUAGE_CODE_TO_NAME",
-    "REWRITE_SYSTEM_INSTRUCTION",
-    "PROMPT_TEMPLATE",
-]
