+++ src/misinformation_simulation/__init__.py
@@ -1,14 +1,13 @@
-from enums.providers import Provider
-
-from .bert_audit_functions import (
+from .audits import (
     consistency_flag,
     nli_pair_scores,
     pretrained_fake_news_detector_prediction,
     read_dataset,
     validate_pair_columns,
 )
-from .simulation_functions import rewrite_news_with_personality
-from .topic_drift_functions import (
+from .enums import DefaultPersonality, Models, Provider
+from .llm import rewrite_news_with_personality
+from .topic_drift import (
     annotate_stdi_for_rewrites,
     annotate_stdi_for_version_chain,
     calculate_stdi,
@@ -17,13 +16,15 @@ from .topic_drift_functions import (
 )
 
 __all__ = [
+    "DefaultPersonality",
+    "Models",
+    "Provider",
     "rewrite_news_with_personality",
     "extract_topic_structure",
     "calculate_stdi",
     "calculate_stdi_chain_metrics",
     "annotate_stdi_for_rewrites",
     "annotate_stdi_for_version_chain",
-    "Provider",
     "read_dataset",
     "validate_pair_columns",
     "consistency_flag",
