+++ src/misinformation_simulation/enums/personalities.py
