+++ src/misinformation_simulation/llm/__init__.py
@@ -1,0 +1,3 @@
+from .rewrite import rewrite_news_with_personality
+
+__all__ = ["rewrite_news_with_personality"]
