+++ src/misinformation_simulation/reporting/__init__.py
@@ -1,0 +1,3 @@
+from .execution_report import append_execution_report, resolve_execution_report_path
+
+__all__ = ["append_execution_report", "resolve_execution_report_path"]
