+++ src/misinformation_simulation/topic_drift/__init__.py
@@ -1,0 +1,16 @@
+from .annotation import annotate_stdi_for_rewrites, annotate_stdi_for_version_chain
+from .extraction import extract_topic_structure
+from .metrics import calculate_stdi, calculate_stdi_chain_metrics
+from .models import TopicRelation, TopicStructure, flatten_topic_structure, topic_structure_to_dict
+
+__all__ = [
+    "TopicRelation",
+    "TopicStructure",
+    "topic_structure_to_dict",
+    "flatten_topic_structure",
+    "extract_topic_structure",
+    "calculate_stdi",
+    "calculate_stdi_chain_metrics",
+    "annotate_stdi_for_rewrites",
+    "annotate_stdi_for_version_chain",
+]
